split everything into components
